+++ src/money.cpp
@@ -78,11 +78,11 @@ std::ostream& budget::operator<<(std::ostream& stream, const money& amount){
     return stream << money_to_string(amount);
 }
 
-budget::money budget::random_money(size_t min, size_t max){
+budget::money budget::random_money(long min, long max){
     static std::random_device rd;
     static std::mt19937_64 engine(rd());
 
-    std::uniform_int_distribution<int> dollars_dist(min, max);
+    std::uniform_int_distribution<long> dollars_dist(min, max);
     std::uniform_int_distribution<int> cents_dist(0, 99);
 
     return {dollars_dist(engine), cents_dist(engine)};
